+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v0.6.0 - 2026-09-20
+
+[Changes since v0.5.0](https://github.com/geewiz/czmq_ada/compare/v0.5.0...v0.6.0)
+
+### Added
+
+- Mandatory ROUTER routing via `Set_Router_Mandatory` and send outcomes via
+  `CZMQ.Messages.Send_Status` ([#19](https://github.com/geewiz/czmq_ada/issues/19)).
+
 ## v0.5.0 - 2026-05-09
 
 [Changes since v0.4.0](https://github.com/geewiz/czmq_ada/compare/v0.4.0...v0.5.0)